momo-backend-ruoyi: 薄荷健康（仿）： 1、用户资料表等五个表等接口开发

+++ momo-modules/momo-health/src/main/java/org/dromara/health/domain/bo/FoodInfoBo.java
@@ -47,27 +47,27 @@ public class FoodInfoBo extends BaseEntity {
      * 热量(kcal)
      */
     @NotNull(message = "热量(kcal)不能为空", groups = { AddGroup.class, EditGroup.class })
-    private Long calories;
+    private Double calories;
 
     /**
      * 蛋白质(g)
      */
-    private Long protein;
+    private Double protein;
 
     /**
      * 脂肪(g)
      */
-    private Long fat;
+    private Double fat;
 
     /**
      * 碳水化合物(g)
      */
-    private Long carbohydrate;
+    private Double carbohydrate;
 
     /**
      * 膳食纤维(g)
      */
-    private Long fiber;
+    private Double fiber;
 
     /**
      * 备注
